% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_results}
\alias{add_results}
\title{Add validation results to the Report object}
\usage{
add_results(data, report)
}
\arguments{
\item{data}{Data that was validated.}

\item{report}{Report object to store validation results.}
}
\description{
This function adds results to validator object with aggregating summary of
  success, error and warning checks. Moreover it parses assertr results attributes and stores
  them inside usable table.
}
