% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{assign_data_line_number}
\alias{assign_data_line_number}
\title{Assign data line number.}
\usage{
assign_data_line_number(data, identifier, file_path)
}
\arguments{
\item{data}{Data.}

\item{identifier}{Identifier.}

\item{file_path}{File path.}
}
\value{
Line number assigned.
}
\description{
Assign data line number.
}
