% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_parsers.R
\name{convert_error_df}
\alias{convert_error_df}
\title{Convert error table column types}
\usage{
convert_error_df(error_df)
}
\arguments{
\item{error_df}{Table consisting assertr error details}
}
\description{
Convert error table column types
}
