% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{create_summary_row}
\alias{create_summary_row}
\title{Create summary table row.}
\usage{
create_summary_row(id, number, color, label)
}
\arguments{
\item{id}{ID.}

\item{number}{Number to display.}

\item{color}{Color of the label.}

\item{label}{Label to display.}
}
\value{
Summary table row.
}
\description{
Create summary table row.
}
