% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{create_validation_results}
\alias{create_validation_results}
\title{Creates data frame from validation results.}
\usage{
create_validation_results(data, errors, warnings, file_path, object_name)
}
\arguments{
\item{data}{Validated data frame.}

\item{errors}{Data frame with validation errors. See \code{get_validations_attribute} function.}

\item{warnings}{Data frame with validation warnings. See \code{get_validations_attribute} function.}

\item{file_path}{Github repo file path to redirect.}

\item{object_name}{Title to display in the report.}
}
\value{
Data frame with validation results.
}
\description{
Creates data frame from validation results.
}
