% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{create_validator}
\alias{create_validator}
\title{Create new validator object}
\usage{
create_validator()
}
\description{
The object returns R6 class environment resposible for storing validation results.
}
