% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{data_validation_report}
\alias{data_validation_report}
\title{Create new validator object}
\usage{
data_validation_report()
}
\description{
The object returns R6 class environment responsible for storing validation results.
}
