% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\docType{data}
\name{datavalidator_constants}
\alias{datavalidator_constants}
\title{Report constants.}
\format{An object of class \code{environment} of length 3.}
\usage{
datavalidator_constants
}
\value{
Report constants.
}
\description{
Report constants.
}
\keyword{datasets}
