% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report_helpers.R
\name{display_results}
\alias{display_results}
\title{Displays results of validations.}
\usage{
display_results(data, repo_path = NULL)
}
\arguments{
\item{data}{Report data.}

\item{repo_path}{Github repo path.}
}
\value{
Validation report.
}
\description{
Displays results of validations.
}
