% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\docType{data}
\name{.validator}
\alias{.validator}
\title{Default validator object}
\format{
An object of class \code{Validator} (inherits from \code{R6}) of length 9.
}
\usage{
.validator
}
\description{
Default validator object
}
\keyword{datasets}
