% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_first_noncall}
\alias{find_first_noncall}
\title{Create a recursive function to find the first non-call object}
\usage{
find_first_noncall(object)
}
\arguments{
\item{object}{A list representing an R expression.}
}
\value{
The first non-call object found in the list representation of an R expression.
}
\description{
This function iteratively dives into the provided list (R expression),
until it finds an object that is not a function call or a complex command.
The [[2]] is used with the argument 2, because in the list representation
of function calls in R, the actual function is the first element, and its arguments
are the subsequent elements. So object 2 generally refers to the first argument
of the function call.
}
\keyword{internal}
