% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML_report_helpers.R
\name{generate_html_report}
\alias{generate_html_report}
\title{Generate HTML report.}
\usage{
generate_html_report(n_passed, n_failed, n_warned, validation_results,
  repo_path)
}
\arguments{
\item{n_passed}{Nr of passed validations}

\item{n_failed}{Nr of failed validations.}

\item{n_warned}{Nr of warnings.}

\item{validation_results}{Data frame with validation results.}

\item{repo_path}{Github repo path.}
}
\value{
HTML validation report.
}
\description{
Generate HTML validation report.
}
