% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\name{generate_report}
\alias{generate_report}
\title{Generate validation report.}
\usage{
generate_report(template, output_dir, output_file = "validation_report.html",
  scripts, repo_path = NULL)
}
\arguments{
\item{template}{Report Rmd template.}

\item{output_dir}{Output directory for HTML report.}

\item{output_file}{Output name for HTML report.}

\item{scripts}{Vector containing paths for the scripts to run inside the template (e.g. data reading, validations). Note that the paths should be relative to the location of the template.}

\item{repo_path}{Github repo path.}
}
\value{
Validation report.
}
\description{
Generate validation report.
}
