% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{get_assert_method}
\alias{get_assert_method}
\title{Match proper method depending on predicate type}
\usage{
get_assert_method(
  predicate,
  method = list(direct = assertr::assert, generator = assertr::insist)
)
}
\arguments{
\item{predicate}{Predicate or predicate generator function.}

\item{methods}{List of two elements named 'direct' and 'generator' that will be returned depending on predicate type.}
}
\description{
Match proper method depending on predicate type
}
