% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_parsers.R
\name{get_assertion_type}
\alias{get_assertion_type}
\title{get assertion type}
\usage{
get_assertion_type(assertion)
}
\arguments{
\item{assertion}{assertion object (check \code{assertr} package for details)}
}
\value{
character with id of assertion: "error", "success", "warning"
}
\description{
get assertion type
}
