% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ast}
\alias{get_ast}
\title{Constructs an Abstract Syntax Tree for an expression}
\usage{
get_ast(exp)
}
\arguments{
\item{exp}{An R expression to be parsed into a list structure.}
}
\value{
A list structure that represents the input R expression.
}
\description{
This function breaks down an R expression into a list structure,
creating a tree-like representation of the code.
}
\keyword{internal}
