% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_predicate_type}
\alias{get_predicate_type}
\title{Detect type of predicate}
\usage{
get_predicate_type(predicate)
}
\arguments{
\item{predicate}{Predicate or predicate generator function.}
}
\value{
String informing about predicate type.
  When "direct" the predicate is directly applied to the column.
  When "generator" the predicate returns predicate computed based on a column.
  When NA, the predicate shouldn't be used at all in the validation.
}
\description{
Detect type of predicate
}
