% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{get_results}
\alias{get_results}
\title{Get validation results}
\usage{
get_results(unnest = FALSE, validator = getDefaultValidator())
}
\arguments{
\item{unnest}{If TRUE, error_df table is unnested. Results with remaining columns duplicated in table.}

\item{validator}{Validator object that stores validation results. See \link{add_results}.}
}
\description{
The response is a list containing information about successful, failed, warning assertions and
  the table stores important information about validation results. Those are:
  \itemize{
    \item table_name - name of validated table
    \item assertion.id - id used for each assertion
    \item description - assertion description
    \item num.violations - number of violations (assertion and column specific)
    \item call - assertion call
    \item message - assertion result message for specific column
    \item type - error, warning or success
    \item error_df - nested table storing details about error or warning result (like vilated indexes and valies)
  }
}
