% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_parsers.R
\name{get_results_number}
\alias{get_results_number}
\title{Get results number}
\usage{
get_results_number(results)
}
\arguments{
\item{results}{assertion results}
}
\value{
table with results number
}
\description{
Get results number
}
