% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report_helpers.R
\name{get_semantic_report_ui}
\alias{get_semantic_report_ui}
\title{Generate HTML report.}
\usage{
get_semantic_report_ui(n_passed, n_failed, n_warned, validation_results)
}
\arguments{
\item{n_passed}{Number of passed validations}

\item{n_failed}{Number of failed validations.}

\item{n_warned}{Number of warnings.}

\item{validation_results}{Data frame with validation results.}

\item{repo_path}{Github repo path.}
}
\value{
HTML validation report.
}
\description{
Generate HTML validation report.
}
