% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{get_validations_attribute}
\alias{get_validations_attribute}
\title{Creates data frame from validation attribute.}
\usage{
get_validations_attribute(data, attribute)
}
\arguments{
\item{data}{Validated data frame.}

\item{attribute}{Attribute name.}
}
\value{
Data frame with validation results.
}
\description{
Creates data frame from validation attribute.
}
