% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_complex_command}
\alias{is_complex_command}
\title{Check if a command is complex, i.e, contains any non-alphanumeric character}
\usage{
is_complex_command(command_string)
}
\arguments{
\item{command_string}{A character string representing the command to be checked.}
}
\value{
Logical value indicating whether the command_string is complex (TRUE) or not (FALSE).
}
\description{
Check if a command is complex, i.e, contains any non-alphanumeric character
}
\keyword{internal}
