% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{make_accordion_container}
\alias{make_accordion_container}
\title{Create a UI accordion container.}
\usage{
make_accordion_container(...)
}
\arguments{
\item{...}{Additional agruments inside accordion container.}
}
\value{
Accordion container.
}
\description{
Create a UI accordion container.
}
