% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report_helpers.R
\name{make_summary_table}
\alias{make_summary_table}
\title{Create summary table.}
\usage{
make_summary_table(n_passes, n_fails, n_warns)
}
\arguments{
\item{n_passes}{Nr of passed validations.}

\item{n_fails}{Nr of failed validations.}

\item{n_warns}{Nr of warnings.}
}
\value{
Summary table.
}
\description{
Create summary table.
}
