% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{make_table_row}
\alias{make_table_row}
\title{Create table row.}
\usage{
make_table_row(title, result, id, data_name, errors, mark)
}
\arguments{
\item{title}{Row title.}

\item{result}{Result to display in row.}

\item{id}{Row id.}

\item{data_name}{Data name.}

\item{errors}{Errors.}

\item{mark}{Mark.}
}
\value{
Table row.
}
\description{
Create table row.
}
