% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\docType{data}
\name{n_failed}
\alias{n_failed}
\title{n_failed..}
\format{An object of class \code{numeric} of length 1.}
\usage{
n_failed
}
\value{
n_failed.
}
\description{
n_failed.
}
\keyword{datasets}
