% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\docType{data}
\name{n_warned}
\alias{n_warned}
\title{n_failed.}
\format{An object of class \code{numeric} of length 1.}
\usage{
n_warned
}
\value{
n_warned.
}
\description{
n_warned.
}
\keyword{datasets}
