% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_parsers.R
\name{parse_errors_to_df}
\alias{parse_errors_to_df}
\title{Parse errors to data.frame}
\usage{
parse_errors_to_df(data)
}
\arguments{
\item{data}{object of assertr error class  (check \code{assertr} package
for details)}
}
\value{
data.frame with errors
}
\description{
Parse errors to data.frame
}
