% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_parsers.R
\name{parse_results_to_df}
\alias{parse_results_to_df}
\title{Parse results to data.frame}
\usage{
parse_results_to_df(data)
}
\arguments{
\item{data}{assertr object  (check \code{assertr} package for details)}
}
\value{
data.frame with successes and errors
}
\description{
Parse results to data.frame
}
