% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_parsers.R
\name{parse_successes_to_df}
\alias{parse_successes_to_df}
\title{Parse successes to data.frame}
\usage{
parse_successes_to_df(data)
}
\arguments{
\item{data}{object of assertr success class  (check \code{assertr} package
for details)}
}
\value{
data.frame with successes
}
\description{
Parse successes to data.frame
}
