% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{prepare_modal_content}
\alias{prepare_modal_content}
\title{Prepare modal content.}
\usage{
prepare_modal_content(error, df_error_head_n)
}
\arguments{
\item{error}{Assertr error.}

\item{df_error_head_n}{Number of rows to display in error table.}
}
\value{
Modal content.
}
\description{
Prepare modal content.
}
\keyword{internal}
