% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{render_raw_report_ui}
\alias{render_raw_report_ui}
\title{Render simple version of report}
\usage{
render_raw_report_ui(n_passed, n_failed, n_warned, validation_results)
}
\arguments{
\item{n_passed}{Number of successful assertions.}

\item{n_failed}{Number of warning assertions.}

\item{n_warned}{Number of violation assertions.}

\item{validation_results}{Validation results table (see \link{get_results}).}
}
\description{
Renders content of simple report version that prints \code{validation_results} table.
}
