% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{render_semantic_report_ui}
\alias{render_semantic_report_ui}
\title{Render semantic version of report}
\usage{
render_semantic_report_ui(
  validation_results,
  success = TRUE,
  warning = TRUE,
  error = TRUE,
  df_error_head_n = 6L
)
}
\arguments{
\item{validation_results}{Validation results table (see \link{get_results}).}

\item{success}{Should success results be presented?}

\item{warning}{Should warning results be presented?}

\item{error}{Should error results be presented?}

\item{df_error_head_n}{Number of rows to display in error table. Works in the same way as \code{head} function.}
}
\description{
Renders content of semantic report version.
}
