% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{result_table}
\alias{result_table}
\title{Create table with results.}
\usage{
result_table(results, errors, data_name, mark)
}
\arguments{
\item{results}{Result to display in table.}

\item{errors}{Errors.}

\item{data_name}{Data name.}

\item{mark}{Mark.}
}
\value{
Table row.
}
\description{
Create table with results.
}
