% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{return_line_number}
\alias{return_line_number}
\title{Return line number.}
\usage{
return_line_number(identifier, file_path, offset = -1)
}
\arguments{
\item{identifier}{Identifier.}

\item{file_path}{File path.}

\item{offset}{Offset.}
}
\value{
Line number.
}
\description{
Return line number.
}
