% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{return_repo_link}
\alias{return_repo_link}
\title{Return repo link.}
\usage{
return_repo_link(repo_path, line_number)
}
\arguments{
\item{repo_path}{Github repo path.}

\item{line_number}{Line number.}
}
\value{
Repo link.
}
\description{
Return repo link.
}
