% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{save_report}
\alias{save_report}
\title{Saving results as a HTML report}
\usage{
save_report(
  validator,
  output_file = "validation_report.html",
  output_dir = getwd(),
  summary = c("error", "warning", "success"),
  ui_constructor = render_semantic_report_ui,
  template = system.file("rmarkdown/templates/standard/skeleton/skeleton.Rmd", package =
    "data.validator")
)
}
\arguments{
\item{validator}{Validator object that stores validation results.}

\item{output_file}{Html file name to write report to.}

\item{output_dir}{Report directory.}

\item{summary}{Vector storing at least one of \code{"error"}, \code{"warning"}, \code{"success"}.
Results with using validation filtered out results table in report.
Moreover, skipping one of the options sets up \code{n_failed}, \code{n_warned} or \code{n_passed}
that is passed to \code{ui_contructor} function as NULL - helpful if you want to distinguish whether
to include related information in report or not (comparing to value 0 that means there were no results
of related type).}

\item{ui_constructor}{Function of four arguments \code{n_failed}, \code{n_warned} or \code{n_passed} and
\code{validation_reults} that generates HTML code or HTML widget that should be used to generate report
content. See \code{custom_report} example.}

\item{template}{Path to Rmd template in which ui_contructor is rendered. See \code{data.validator} rmarkdown
template to see basic construction - the one is used as a default template.}
}
\description{
Saving results as a HTML report
}
