% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{save_summary}
\alias{save_summary}
\title{Save simple validation summary in text file}
\usage{
save_summary(validator, file_name = "validation_log.txt")
}
\arguments{
\item{validator}{Validator object that stores validation results.}

\item{file_name}{Name of the resulting file (including extension).}
}
\description{
Saves \code{print(validator)} output inside text file.
}
