% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{segment}
\alias{segment}
\title{Create a UI segment element.}
\usage{
segment(title, ..., color = "blue")
}
\arguments{
\item{title}{Title of the segment.}

\item{...}{Additional agruments inside sagment.}

\item{color}{Color of the segment.}
}
\value{
Segment.
}
\description{
Create a UI segment element.
}
