% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate}
\alias{validate}
\title{Validate}
\usage{
validate(data, name, description = NULL)
}
\arguments{
\item{data}{data.frame or tibble to test}

\item{name}{name of validation object (will go to report)}

\item{description}{description of validation object (will go to report)}
}
\description{
Prepared data for validation and generating report.
}
