% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_cols}
\alias{validate_cols}
\title{Validation on columns}
\usage{
validate_cols(
  data,
  predicate,
  ...,
  obligatory = FALSE,
  description = NA,
  skip_chain_opts = FALSE,
  success_fun = assertr::success_append,
  error_fun = assertr::error_append,
  defect_fun = assertr::defect_append
)
}
\arguments{
\item{data}{data.frame or tibble to test}

\item{predicate}{Predicate functions from asserts package, e.g. \code{in_set},
\code{within_bounds}, etc.}

\item{...}{other arguments to predicate function}

\item{obligatory}{flag setting assertion to mandatory}

\item{description}{character with description of assertion}

\item{skip_chain_opts}{flag to skip chain options}

\item{success_fun}{function that is called after success}

\item{error_fun}{function that is called after error}

\item{defect_fun}{function that is called after defect}
}
\description{
Validation on columns
}
\seealso{
validate_if validate_rows
}
