% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\docType{class}
\name{Validator}
\alias{Validator}
\title{Class providing object with methods for simple data validation reports.}
\format{\code{\link{R6Class}} object.}
\usage{
Validator
}
\value{
Object of \code{\link{R6Class}} with methods for simple data validation reports.
}
\description{
Class providing object with methods for simple data validation reports.
}
\section{Fields}{

\describe{
\item{\code{repo_path}}{Github repository url of a project.}
}}

\section{Methods}{

\describe{
\item{\code{add_validations(data, file_path = NULL)}}{This method adds \code{assertr} validation results to the report.}
\item{\code{get_validations()}}{This method returns list of current validations.}
\item{\code{generate_html_report()}}{This method generates \code{HTML} validation report.}
\item{\code{save_log(output_path = "validation_log")}}{This method saves report log into \code{output_path}}}
}

\examples{
validator <- Validator$new()
}
\keyword{data}
