% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{add_results}
\alias{add_results}
\title{Add validation results to validator object}
\usage{
add_results(data, validator = getDefaultValidator(), name = NULL)
}
\arguments{
\item{data}{Data that was validated.}

\item{validator}{Validator object to store results to created with \link{create_validator}.}

\item{name}{Optional name for data that was validated.
By default the name of first object used in validation chain.}
}
\description{
This function adds results to validator object with aggregating summary of
  success, error and warning checks. Moreover it parses assertr results attributes and stores
  them inside usable table.
}
