% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_costructors.R
\name{display_results}
\alias{display_results}
\title{Displays results of validations.}
\usage{
display_results(data, n_passed, n_failed, n_warned)
}
\arguments{
\item{data}{Report data.}
}
\value{
Validation report.
}
\description{
Displays results of validations.
}
