% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{display_results}
\alias{display_results}
\title{Displays results of validations.}
\usage{
display_results(data, make_report, repo_path = NULL)
}
\arguments{
\item{data}{Report data.}

\item{make_report}{Make report.}

\item{repo_path}{Github repo path.}
}
\value{
Validation report.
}
\description{
Displays results of validations.
}
