% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{display_results}
\alias{display_results}
\title{Displays results of validations.}
\usage{
display_results(data, n_passes, n_fails, n_warns)
}
\arguments{
\item{data}{Report data.}

\item{n_passes}{Number of successful assertions.}

\item{n_fails}{Number of warning assertions.}

\item{n_warns}{Number of violation assertions.}
}
\value{
Validation report.
}
\description{
Displays results of validations.
}
