% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_first_name}
\alias{get_first_name}
\title{Get name of the data frame based on active frames}
\usage{
get_first_name()
}
\value{
deparsed chain part
}
\description{
Get name of the data frame based on active frames
}
\keyword{internal}
