% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_first_name}
\alias{get_first_name}
\title{Get first name of the data frame}
\usage{
get_first_name(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
deparesed chain part
}
\description{
Get first name of the data frame
}
