% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{get_semantic_report_ui}
\alias{get_semantic_report_ui}
\title{Generate HTML report.}
\usage{
get_semantic_report_ui(n_passes, n_fails, n_warns, validation_results)
}
\arguments{
\item{n_passes}{Number of passed validations}

\item{n_fails}{Number of failed validations.}

\item{n_warns}{Number of warnings.}

\item{validation_results}{Data frame with validation results.}
}
\value{
HTML validation report.
}
\description{
Generate HTML validation report.
}
