% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{make_accordion_element}
\alias{make_accordion_element}
\title{Create a UI accordion element.}
\usage{
make_accordion_element(title, content, active = FALSE)
}
\arguments{
\item{title}{Title of the accordion.}

\item{content}{Content of the accordion.}

\item{active}{Is active?}
}
\value{
Accordion.
}
\description{
Create a UI accordion element.
}
