% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{make_table_row}
\alias{make_table_row}
\title{Create table row.}
\usage{
make_table_row(results, type, mark)
}
\arguments{
\item{results}{Results to display in a row.}

\item{type}{Result type.}

\item{mark}{Icon to display.}
}
\value{
Table row.
}
\description{
Create table row.
}
