% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{prepare_modal_content}
\alias{prepare_modal_content}
\title{Prepare modal content.}
\usage{
prepare_modal_content(error)
}
\arguments{
\item{error}{Error.}
}
\value{
Modal content.
}
\description{
Prepare modal content.
}
