% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{prepare_modal_content}
\alias{prepare_modal_content}
\title{Prepare modal content.}
\usage{
prepare_modal_content(error)
}
\arguments{
\item{error}{Assertr error.}
}
\value{
Modal content.
}
\description{
Prepare modal content.
}
