% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{result_table}
\alias{result_table}
\title{Create table with results.}
\usage{
result_table(results, type, mark)
}
\arguments{
\item{results}{Result to display in table.}

\item{type}{Result type.}

\item{mark}{Icon to display.}
}
\value{
Table row.
}
\description{
Create table with results.
}
