% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{save_results}
\alias{save_results}
\title{Saving results table to external file}
\usage{
save_results(report, file_name = "results.csv", method = utils::write.csv, ...)
}
\arguments{
\item{report}{Report object that stores validation results. See \link{get_results}.}

\item{file_name}{Name of the resulting file (including extension).}

\item{method}{Function that should be used to save results table (write.csv default)
The function passed to \code{method} should have 'x' and 'file' arguments. Functions with
different arguments can be passed by creating a wrapper function for it.
See example save_results_methods.}

\item{...}{Remaining parameters passed to \code{method}.}
}
\description{
Saving results table to external file
}
