% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{save_results}
\alias{save_results}
\title{Saving results table to external file}
\usage{
save_results(
  file_name = "results.csv",
  method = utils::write.csv,
  validator = getDefaultValidator(),
  ...
)
}
\arguments{
\item{file_name}{Name of the resulting file (including extension).}

\item{method}{Function that should be used to save results table (write.csv default).}

\item{validator}{Validator object that stores validation results. See \link{get_results}.}

\item{...}{Remaining parameters passed to \code{method}.}
}
\description{
Saving results table to external file
}
