% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{save_summary}
\alias{save_summary}
\title{Save simple validation summary in text file}
\usage{
save_summary(
  file_name = "validation_log.txt",
  success = TRUE,
  warning = TRUE,
  error = TRUE,
  validator = getDefaultValidator()
)
}
\arguments{
\item{file_name}{Name of the resulting file (including extension).}

\item{success}{Should success results be presented?}

\item{warning}{Should warning results be presented?}

\item{error}{Should error results be presented?}

\item{validator}{Validator object that stores validation results.}
}
\description{
Saves \code{print(validator)} output inside text file.
}
