% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_report_constructors.R
\name{segment}
\alias{segment}
\title{Create a UI segment element.}
\usage{
segment(title, ...)
}
\arguments{
\item{title}{Title of the segment.}

\item{...}{Additional agruments inside segment.}
}
\value{
Segment.
}
\description{
Create a UI segment element.
}
