% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate}
\alias{validate}
\title{Prepare data for validation chain}
\usage{
validate(data, name, description = NULL)
}
\arguments{
\item{data}{data.frame or tibble to test}

\item{name}{name of validation object (will be displayed in the report)}

\item{description}{description of validation object (will be displayed in the report)}
}
\description{
Prepare data for validation and generating report.
The function prepares data for chain validation and ensures all the validation results are
gathered correctly. The function also attaches additional information to the data
(name and description) that is then displayed in validation report.
}
