% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_if}
\alias{validate_if}
\title{"If" Validation}
\usage{
validate_if(
  data,
  expr,
  description = NA,
  obligatory = FALSE,
  skip_chain_opts = FALSE,
  success_fun = assertr::success_append,
  error_fun = assertr::error_append,
  defect_fun = assertr::defect_append
)
}
\arguments{
\item{data}{data.frame or tibble to test}

\item{expr}{expression to test for, e.g. \code{within_n_sds(3)}}

\item{description}{character with description of assertion}

\item{obligatory}{flag setting assertion to mandatory}

\item{skip_chain_opts}{flag to skip chain options}

\item{success_fun}{function that is called after success}

\item{error_fun}{function that is called after error}

\item{defect_fun}{function that is called after defect}
}
\description{
Check if an assertion is met.
}
\seealso{
validate_cols validate_rows
}
